requirements.txt
tensorflow
pandas
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyyaml
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
flask
Flask-Cors
-e .
